Case-sensitive FSSpec
This keeps Pomme from trawling through a directory looking for case-insensitive matches, which can be problematic if it stumbles upon a file name with encoding errors. A Cro-Mag Rally user on Windows 10 got an uncaught exception that was thrown while the game was trying to find its prefs because some file in their %APPDATA% had a weird encoding. This should be solved by Pomme having switched to u8string for anything that touches the filesystem, but let's just drop case insensitive FSSpecs altogether.